chore: audit-readiness updates across SDK, vault, DCAP, and CLI
Carry the Digital Seal evidence-bundle work through the Python and TypeScript
SDKs, the vault keeper, the Intel DCAP verifier, and the seal CLI, with the
accompanying deployment-validation script updates.

+++ crates/core/src/pillars/quadratic_burn.rs
@@ -927,7 +927,7 @@ mod tests {
         for i in 0..20 {
             tracker.record_block(BlockGasUsage {
                 block_number: i,
-                gas_used: 15_000_000 + (i as u64 * 500_000), // Increasing load
+                gas_used: 15_000_000 + (i * 500_000), // Increasing load
                 gas_limit: 30_000_000,
                 timestamp: 1000 + i,
                 transaction_count: 100,
